+++ cmake/headerlist.cmake
@@ -1,0 +1,7 @@
+set(headers ${headers}
+        src/Logging.h
+        src/Papyrus.h
+        src/Settings.h
+        src/UI.h
+        src/Utils.h
+)