+++ cmake/sourcelist.cmake
@@ -1,0 +1,8 @@
+set(sources ${sources}
+        src/Logging.cpp
+        src/Papyrus.cpp
+        src/Settings.cpp
+        src/UI.cpp
+        src/Utils.cpp
+        src/Main.cpp
+)