+++ src/Logging.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace kxWhereAreYou::Logging {
+    void Setup();
+}