+++ src/PCH.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <RE/Skyrim.h>
+#include <SKSE/SKSE.h>
+
+#include <algorithm>
+
+namespace logger = SKSE::log;
+
+using namespace std::literals;