+++ src/Papyrus.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace kxWhereAreYou::Papyrus {
+    void Setup();
+}